+++ webclient/src/components/LandingPageSections/Navbar/Navbar.js
@@ -1,18 +1,9 @@
-import React, { Fragment } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import "./Navbar.scss";
 import { NavLink } from "react-router-dom";
-import Modal from "../../UI/Modal/Toggler";
-import Login from "../../Auth/Login";
-import Register from "../../Auth/Register";
 
-const navbar = props => {
-  const { error } = props;
-  let errorMessage = null;
-
-  if (error) {
-    errorMessage = error.response.data.error.message;
-  }
+const navbar = ({ toggleForms }) => {
   return (
     <nav className="Landing-Navbar">
       <div className="nav-wrapper Container">
@@ -27,30 +18,9 @@ const navbar = props => {
             <NavLink to="/">Features</NavLink>
           </li>
           <li>
-            <Modal
-              state={props.state}
-              onClose={props.onClose}
-              onSubmit={props.onSubmit}
-              buttonClasses="AuthButton"
-              modalClasses="AuthModal"
-              title="Join Us"
-            >
-              <Fragment>
-                <Login
-                  submitHandler={props.onSubmit}
-                  onChangeHandler={props.onChange}
-                  {...props.state.login}
-                />
-                {errorMessage ? (
-                  <p style={{ color: "red" }}>{errorMessage}</p>
-                ) : null}
-                <Register
-                  submitHandler={props.onSubmit}
-                  onChangeHandler={props.onChange}
-                  {...props.state.register}
-                />
-              </Fragment>
-            </Modal>
+            <button className="AuthButton" onClick={toggleForms}>
+              Register / Log in
+            </button>
           </li>
         </ul>
       </div>
@@ -59,10 +29,7 @@ const navbar = props => {
 };
 
 navbar.propTypes = {
-  state: PropTypes.object.isRequired,
-  onSubmit: PropTypes.func.isRequired,
-  onChange: PropTypes.func.isRequired,
-  error: PropTypes.object
+  toggleForms: PropTypes.func.isRequired
 };
 
 export default navbar;