+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.4] - 2019-07-04
+
+### Added
+
+- Added form validation in both frontend and backend
+- Added colors for invalid form inputs
+
+### Changed
+
+- Improved component structure to fit the form validation changes
+- Moved the authentication actions from redux to the LandingPage component
+- Added a close button in the authentication modal
+- Made the colors in the landing page more consistent (replaced Material UI's default color)
+- Added prop-types to components
+- Cleaned up console.logs and commented out code
+- Minor style improvements
+
 ## [1.5.3] - 2019-03-28
 
 ### Changed